requirements.txt
blinker==1.9.0
click==8.1.8
coverage==7.8.0
Flask==3.1.0
flask-cors==5.0.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.5.0
PyJWT==2.10.1
pytest==8.3.5
pytest-cov==6.1.1
Werkzeug==3.1.3

